+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - **Added** app instance events
     - `on` adds a listener given an event name
     - `emit` call an event
+- **Added** `autoDraw` property to Doz constructor, by default it's false
+- **Added** `draw` method to Doz instance
 
 ## [1.5.0] - 2018-09-09
 - **Added** class pattern, now it's possible create component with class ES6 syntax